+++ src/authentication-flows/index.ts
@@ -1,4 +1,4 @@
-export * from "./passwordless";
+export * from "./ticket";
 export * from "./silent";
 export * from "./social";
 export * from "./universal";
